art-exhibits/requirements.txt
# Art Exhibits Calendar Sync Dependencies

# Google API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional: For JavaScript-heavy sites
# playwright>=1.40.0
# selenium>=4.15.0
